llm_service/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
openai